Started a huge refactor on the grid
Started refactoring the grid because the old system was really hard to
work with and the data structure I was using didn't really make sense
anymore. Tiles can move around, there is no merge function and a new
tile doesn't appear when a move is made, but the tiles are animated when
they move, so YAY! =D Also, there is still a key-spamming bug with the
tile movement, need to remember to fix that.