Fixed a bug and refactored keypress handling
Fixed the bug where the window scrolls every time the 'up' or 'down' is
pressed in the game.

Moved the whole keypress queueing system from script.js to Game.js. This
is so that I can eventually queue inputs from the automated algorithm
that way automated gameplay can also be animated. Also because it just
makes more sense in here. Automated gameplay works with animation right
now, but it needs to be tested a little more.